+++ books/books_manager.rb
