+++ people/people_manager.rb
