+++ people/person.rb
