+++ rentals/rentals_manager.rb
